+++ src/refactoring/hooks/admin/useNewCoupon.ts
@@ -1,0 +1,27 @@
+import { useState } from "react";
+import { Coupon } from "../../../types";
+
+export const useNewCoupon = (onCouponAdd: (coupon: Coupon) => void) => {
+  const [newCoupon, setNewCoupon] = useState<Coupon>({
+    name: "",
+    code: "",
+    discountType: "percentage",
+    discountValue: 0,
+  });
+
+  const handleAddCoupon = () => {
+    onCouponAdd(newCoupon);
+    setNewCoupon({
+      name: "",
+      code: "",
+      discountType: "percentage",
+      discountValue: 0,
+    });
+  };
+
+  return {
+    newCoupon,
+    setNewCoupon,
+    handleAddCoupon,
+  };
+};
